+++ CHANGELOG.md
@@ -4,4 +4,4 @@ All notable changes to this project will be documented in this file.
 
 ## `bullet 2.2.0`
 - Add optional return indexes for `Bullet`, `Check`, and `Scrollbar`.
-- Fix `YesNo` keyhandler error (Issue #50) 
+- Fix `YesNo` keyhandler error (Issue #50)